Recursive variant expansion
Summary:
Problem running blubber cfg.yml test with this config:

```
base: nodejs-slim
variants:
  build:
    base: nodejs-devel
  development:
    includes: [build]
    entrypoint: [npm, start]
  test:
    includes: [development]
    entrypoint: [npm, test]
```

yields a Dockerfile with `FROM nodejs-slim`; however, I expected that
the base would be `nodejs-devel` since `test` inherits from
`development` which inherits from `build`.

In order for this to work as expected we have to recursively expand
variants as in this patch.

Reviewers: dduvall, #release-engineering-team

Reviewed By: dduvall, #release-engineering-team

Tags: #release-engineering-team

Differential Revision: https://phabricator.wikimedia.org/D773