Switch to github import path for testify
Summary:
Debian only provides a package for the GitHub version so it's either
this or we'd have to maintain a patch to change the import paths at
packaging, a moving target for sure.

Since it's only a development dependency, we're not too worried about
versioning it and this is a much more simple solution.

Refs T175609

Test Plan: Run `go get ./... && go test ./...`.

Reviewers: thcipriani, #release-engineering-team

Reviewed By: thcipriani, #release-engineering-team

Tags: #release-engineering-team

Maniphest Tasks: T175609

Differential Revision: https://phabricator.wikimedia.org/D782