Provide a `runs.insecurely` to be used with test variants
Summary:
Use cases involving running of test suites and doc generation require
more liberal ownership and read/write permission to application files.
When `runs.insecurely` is set to `true`, the effective runtime user will
be `lives.as`, the same user that owns the application files and
installed dependencies.

D999 is a complement to this change to allow restrictions on this and
other potentially sensitive configuration.

Depends on D999, D1002

Test Plan: Run `go test ./...`.

Reviewers: thcipriani, mmodell, hashar, #release-engineering-team, demon

Reviewed By: thcipriani, #release-engineering-team

Tags: #release-engineering-team

Differential Revision: https://phabricator.wikimedia.org/D1003