Install `node_modules` to application directory
The original design decision to install `node_modules` to a central
location outside the application directory was for a purported
development use case not yet realized. It has led to confusion and
unintended failures, and is just generally weird.

Let's right this wrong before it fossilizes into debt, like a hysterical
raisin in the sun.

Bug: T204591
Change-Id: I2e4797833471596b10b6b4a2705055050ca98ac6