Include a basic go.mod that declares the package name
Work on Go Modules has been moving steadily with basic support in 1.11
and more mature support in the recently released 1.12. Eventually we
should look at using it for dependency management (away from `dep`) but
let's start by simply declaring the package name in a new `go.mod` to
allow building of our project outside the hell that is `GOPATH`.

Change-Id: I90b9f35194afe5538dce8db67b2b7c2d551bfaff