python.go: Add support for Poetry package manager
Poetry (<https://python-poetry.org/>) is a packaging and dependency
management tool for Python. Poetry creates a fully versioned dependency
tree for your project. The poetry.lock file tracks the exact versions of
each library and its dependencies for repeatable installs.

Usage example:
```
variants:
  build:
    python:
      version: python3
      poetry:
        version: ==1.0.10
      requirements: [pyproject.toml, poetry.lock]
  test:
    includes: [build]
    python:
      poetry:
        devel: true
    copies: [local]
  production:
    includes: [build]
    copies: [local]
```

python.poetry.version must be a simplified Python package version
specification. A new "pypkgver" validator is provided to ensure that the
value is usable for a `pip install poetry...` command. Typically it
would be reasonable to pin the version of Poetry in use to an exact
version as show in the sample above, but it may be desirable to use
other more complex constraints such as ">=1.0,!=1.0.3,<2.0".

python.poetry.devel is a boolean flag indicating whether or not to
install development dependencies in the Poetry managed venv. It defaults
to false.

When python.poetry.version is non-empty various changes will be made to
the generated Dockerfile.
; PhasePrivileged
: A version of Poetry matching the python.poetry.version constraint
  will be installed using pip.
: POETRY_VIRTUALENVS_PATH will be set in the environment.
; PhasePreInstall
: PIP_WHEEL_DIR and PIP_FIND_LINKS will not be set in the environment.
: `poetry install --no-root --no-dev` will be used to create and
  populate a venv within POETRY_VIRTUALENVS_PATH.
: If python.poetry.devel is true, the `--no-dev` flag will be omitted.
; PhasePostInstall
: PYTHONPATH and PATH will not be set in the environment.

Change-Id: I33b356ff90983f8b8d5b76003851db139d97fe2e