apt.go: Add ability to target releases
In order to allow packages from backports to be downloaded, the ability
to target a release when running apt-get install is required.

apt.packages now takes the previously defined list of string or a map
of targets and packages to install. When using a map, in order to
install the packages for the base image, the 'default' key should be
used:

packages:
  default: ["libfoo", "libbar"]
  buster-backports: ["libbaaz"]

Bug: T272759
Change-Id: I435c47794ffbc0264e0a440bfbd23e11570645d1