python: Pin pip package to <21 for Python 2
Pip 21 has dropped support for Python 2 but is still strangely
selected for installation when running setuptools with a python2 binary.
Let's pin pip when a `version` is specified that starts with "python2".

Note this will require python 2 users to explicitly specify `version:
python2` even if their default `python` binary in the base image is for
python 2.

Bug: T273793
Change-Id: Id7d4315e345657ce7ac2efdd0b065d693e47b2f5