python: upgrade pip before installing requirements
A `python` config block already triggered installing pip via
easy_install which is pretty specific to python2 era usage. It then used
pip to install/update pypi packages for setuptools, wheel, and tox. This
change adds the pip package to that set of pypi install/update packages.
This is useful to ensure that the resulting image is capable of
installing manylinux2014 formatted wheels (pip >= 19.3). Debian Buster's
python3-pip package is pip 18.1 which is too old to look for
manylinux2014 wheels when installing packages such as cryptography which
have stopped publishing wheels with the older manylinux1 and
manylinux2010 platform tags.

Existing pinning of pip to a version less than 21.0 is preserved for
those poor, unloved python2 projects that linger in an ever colder and
darker world hostile to their kind.

Bug: T274435
Change-Id: I81f0af69dbd4f9202bc9736faf653813c81fc0ad