apt: Implement merging `apt.proxies` config
The `apt.proxies` field is now merged by appending second set of
proxies to the original.

Example:

    apt:
      proxies:
        - url: http://proxy.example:8080
    variants:
      foo:
        apt:
          proxies:
            - url: http://proxy.example:8081
              source: http://security.debian.org

In the above example, the effective `apt.proxies` configuration for the
`foo` variant would include both the base `proxy.example:8080` proxy and
the `proxy.example:8081` proxy specific to the `security.debian.org`
source.

Bug: T277109
Change-Id: I64bad1eb1e6fb6712a5de178be3aa273936a2df5