Ensure dependencies from `includes` are present in the CopiesDepGraph
Under use cases where a user includes a previous variant that contains a
number of `copies`, the dependencies resulting between the `includes`
and `copies` should be aggregately transitive.

The following example illustrates the use case.

    version: v4
    base: foo
    variants:
      one: {}
      two:
        copies: [one]
      three:
        includes: [two]

If one were to compile variant `three`, it would rightly contain `COPY
--from=one` instructions. However, without this change, no `one` Docker
stage would be included in the output and the instructions would fail
when built by Docker.

This change fixes that behavior in `ExpandIncludesAndCopies` by waiting
to add `copies` dependencies to the CopiesDepGraph until after the
variant is expanded via `ExpandVariant` and `copies` for all `includes`
are present in the variant.

Bug: T289880
Change-Id: I2af40e352e8f6208ca914c6d0d0b334e8b691133