apt: Support user defined APT sources
Provide configuration for additional APT sources, allowing users to
specify a repo URL, distribution/release name, and a number of
components.

Note that the source is not marked as trusted and so it must be signed
and the key must be present in the keyring of the base image. This
constraint should protect us from use of arbitrary package sources and
root-installed packages.

Bug: T289367
Change-Id: I3046e755abde16beb7ec8ec5abe5bb48763b93bb