feature: build-time arguments for lives & runs user config
Generate `ARG` instructions with default values taken from the Blubber
config for the username, UID, and GID used by the build-time and runtime
stages of the container build. This allows us to use symbolic names in
generated instructions which can be more readable than the raw values.
It also allows for the possibility of build time changes to be made to
these values using tools such as docker-compose which can enable
creation of a read-write local development environment built from
Blubber defined containers.

Bug: T296046
Change-Id: I30599f3a1351771212a4dd4f9170fb0a96bbe0bf