Add make blubber-buildkit-docker
The Make target is intended to build a local image of our buildkit under
the name `local/blubber-buildkit`. The resulting image can then be
referenced in a local `.pipeline/blubber.yaml` file with:

 # syntax = local/blubber-buildkit

That lets one verify a change to the buildkit locally against a real
world definition of an image.

Change-Id: If724779be4e61f819e3d2657f61be03f437f1a2d