pipeline: Only publish blubber-buildkit when new git tags are pushed
Building and publishing the blubber-buildkit image for every merge,
regardless of whether the version has been incremented, leads to
overwriting of existing version tags with new behavior.

Let's limit the publishing of new blubber-buildkit images to when new
git tags are pushed, adopting a GitOps-y workflow for releases.

Build blubber-buildkit during the rehearse pipeline to ensure it builds
correctly.

Change-Id: I8400be0ea4625781832ac3ecd9e08465de99da03