buildkit: Do not require a .dockerignore file
The `.dockerignore` file is read unconditionally even though it should
be an optional file.

Refactored `readFileFromLocal` to accept a `required` argument. If the
file is not required, the path is stat'ed and if that fails, an empty
byte slice is returned with error `nil`. If the file is required, the
extra stat is skipped.

Follows up 64d181191e0da143bcb04db3c99acb69526561e8

Bug: T320988
Change-Id: Ic0c4dcdd4c772d6e3b1470f0c65cd03780391f64