buildkit: Refactor multi-platform build process
Introduced the `buildResult` type that represents the result of a single
image build. It contains a reference to the resulting image, the image
configuration, extra build info, the target platform, platform export
information, and whether the single result is part of a larger
multi-platform build.

Factored out the call to `CompileToLLB` and the LLB solve (image build),
and image config serialization into a `buildImage` function that returns
a new `buildResult`. The `buildResult` can then be used to add/set the
correct meta data and reference to the final aggregate result.

Follows up 43ec449dd7c39b4f6566730f5872280dff4572de