buildkit: Support building for multiple target platforms
Accept multiple values for `platform`, dispatching platform specific
builds in parallel, and aggregating the result as a manifest list (aka
"fat" manifest) that points to references for each target platform.

This feature brings Blubber's BuildKit frontend into parity with the
Docker frontend with regards to multi-platform support. Users can build
with either `docker buildx build --platform ...` or `buildctl --opt
platform=...` and publish multi-platform manifests with a single
command.

Once this feature is merged and a new blubber image is published, we can
add multiple platforms to this project's `.gitlab-ci.yml` and finally
publish a blubber frontend (under a single tag) that can run on multiple
platforms.

Bug: T318866