buildkit: Specify build platforms based on that of the workers
Without an explicitly provided build platform, `dockerfile2llb` seems to
infer one from the target platforms. This is not desirable behavior when
the platform doesn't match that of the worker, as it will have to be
emulated which may not even be possible in all build environments
(requires a working QEMU) or at the very least very slow.

Let's guarantee that the build platform will match that of the workers
or the buildkitd host (they should be the same in our setup).

Note this is what the dockerfile frontend does when it calls out to
`dockerfile2llb` and the implementation was copied from there.

Bug: T318866