feat: append pythonpath for python
Append existing PYTHONPATH to the predefined pythonpath. This allows users to manipulate the PYTHONPATH as they wish by setting the environment variable PYTHONPATH. This enables a common practice in python applications where the root directory of an application is added to the PYTHONPATH in order to allow imports of custom modules that are not part of a package.