Provide examples as discrete executable acceptance tests
Rewrote most of the examples as acceptance tests in Cucumber format
using Godog for the step implementations.

These tests require the following environment variables to be set and
certain resources to be available:

  * BLUBBER_RUN_EXAMPLES: Instruct ./examples_test.go to run the tests.
  * BUILDKIT_HOST: A functioning buildkitd at this host/socket address.
  * BLUBBER_TEST_IMAGE: A previously built Blubber frontend image ref to
    test against.

Note a number of gofmt errors were corrected due to the Go version
upgrade which came with new formatting rules.

Rewrote sections of the README.md to point folks to the new examples.

Bug: T338160