Provide JSON Schema for Blubber configuration
Include a `api/config.schema.json` JSON Schema document to serve both as
documentation and a base layer of validation.

Refactor `config.ReadConfig` to validate against the schema.

Provide a Makefile target for generating documentation from the schema
and include a CI job that ensures the documentation was generated prior
to code review.

Change-Id: Iba35bd2c7bac3568dd36df83a6c1aab302ddbddc