examples: Describe how defining an entrypoint clears default arguments
Moved the label scenarios to a new "minutiae" feature file. Although
they are important behaviors, these scenarios are not relevant to most
users, so let's keep them separate from the scenarios for core
features.

Added an expectation to one of the basic scenarios that defining an
entrypoint also clears the default arguments.

Change-Id: I1c607759e034c9cdbed678a80e62a88afa363e82